Move verkkokauppa to integrations
