+++ tests/test_external_services/test_verkkokauppa/test_merchant_requests.py
@@ -6,14 +6,14 @@ from typing import Any
 import pytest
 from django.conf import settings
 
-from tilavarauspalvelu.utils.verkkokauppa.merchants.exceptions import CreateMerchantError, UpdateMerchantError
-from tilavarauspalvelu.utils.verkkokauppa.merchants.types import (
+from tilavarauspalvelu.integrations.verkkokauppa.merchants.exceptions import CreateMerchantError, UpdateMerchantError
+from tilavarauspalvelu.integrations.verkkokauppa.merchants.types import (
     CreateMerchantParams,
     Merchant,
     MerchantInfo,
     UpdateMerchantParams,
 )
-from tilavarauspalvelu.utils.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
+from tilavarauspalvelu.integrations.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
 from utils.sentry import SentryLogger
 
 from tests.helpers import patch_method
