+++ tests/test_external_services/test_verkkokauppa/test_order_types.py
@@ -8,8 +8,8 @@ from typing import Any
 import pytest
 from django.conf import settings
 
-from tilavarauspalvelu.utils.verkkokauppa.order.exceptions import ParseOrderError
-from tilavarauspalvelu.utils.verkkokauppa.order.types import (
+from tilavarauspalvelu.integrations.verkkokauppa.order.exceptions import ParseOrderError
+from tilavarauspalvelu.integrations.verkkokauppa.order.types import (
     CreateOrderParams,
     Order,
     OrderCustomer,
