+++ tests/test_external_services/test_verkkokauppa/test_payment_requests.py
@@ -9,9 +9,9 @@ import pytest
 from django.conf import settings
 from requests import Timeout
 
-from tilavarauspalvelu.utils.verkkokauppa.payment.exceptions import GetPaymentError, RefundPaymentError
-from tilavarauspalvelu.utils.verkkokauppa.payment.types import Payment, Refund, RefundStatusResult
-from tilavarauspalvelu.utils.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
+from tilavarauspalvelu.integrations.verkkokauppa.payment.exceptions import GetPaymentError, RefundPaymentError
+from tilavarauspalvelu.integrations.verkkokauppa.payment.types import Payment, Refund, RefundStatusResult
+from tilavarauspalvelu.integrations.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
 from utils.sentry import SentryLogger
 
 from tests.helpers import patch_method
