+++ tests/test_external_services/test_verkkokauppa/test_refund_from_json.py
@@ -6,8 +6,8 @@ import uuid
 import pytest
 from django.conf import settings
 
-from tilavarauspalvelu.utils.verkkokauppa.payment.exceptions import ParseRefundError
-from tilavarauspalvelu.utils.verkkokauppa.payment.types import Refund
+from tilavarauspalvelu.integrations.verkkokauppa.payment.exceptions import ParseRefundError
+from tilavarauspalvelu.integrations.verkkokauppa.payment.types import Refund
 from utils.sentry import SentryLogger
 
 from tests.helpers import patch_method
