+++ tests/test_external_services/test_verkkokauppa/test_tasks.py
@@ -5,8 +5,8 @@ from unittest import mock
 
 import pytest
 
+from tilavarauspalvelu.integrations.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
 from tilavarauspalvelu.tasks import refund_paid_reservation_task
-from tilavarauspalvelu.utils.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
 
 from tests.factories import PaymentOrderFactory, ReservationFactory
 from tests.helpers import patch_method
