+++ tests/test_external_services/test_verkkokauppa/test_webhooks/helpers.py
@@ -6,7 +6,12 @@ from decimal import Decimal
 
 from django.utils.timezone import get_default_timezone
 
-from tilavarauspalvelu.utils.verkkokauppa.payment.types import Payment, PaymentStatus, RefundStatus, RefundStatusResult
+from tilavarauspalvelu.integrations.verkkokauppa.payment.types import (
+    Payment,
+    PaymentStatus,
+    RefundStatus,
+    RefundStatusResult,
+)
 
 
 def get_mock_order_payment_api(remote_id: uuid.UUID, payment_id: uuid.UUID, status: str = "") -> Payment:
