+++ tests/test_external_services/test_verkkokauppa/test_webhooks/test_order_cancel_webhooks.py
@@ -9,9 +9,9 @@ from django.urls import reverse
 from django.utils.timezone import get_default_timezone
 
 from tilavarauspalvelu.enums import OrderStatus
-from tilavarauspalvelu.utils.verkkokauppa.order.exceptions import GetOrderError
-from tilavarauspalvelu.utils.verkkokauppa.order.types import Order
-from tilavarauspalvelu.utils.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
+from tilavarauspalvelu.integrations.verkkokauppa.order.exceptions import GetOrderError
+from tilavarauspalvelu.integrations.verkkokauppa.order.types import Order
+from tilavarauspalvelu.integrations.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
 from utils.sentry import SentryLogger
 
 from tests.factories import PaymentOrderFactory
