+++ tests/test_graphql_api/test_reservation/test_delete.py
@@ -7,10 +7,10 @@ from django.test import override_settings
 
 from tilavarauspalvelu.enums import OrderStatus, ReservationStateChoice
 from tilavarauspalvelu.integrations.email.main import EmailService
+from tilavarauspalvelu.integrations.verkkokauppa.order.exceptions import CancelOrderError
+from tilavarauspalvelu.integrations.verkkokauppa.payment.types import PaymentStatus
+from tilavarauspalvelu.integrations.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
 from tilavarauspalvelu.models import Reservation
-from tilavarauspalvelu.utils.verkkokauppa.order.exceptions import CancelOrderError
-from tilavarauspalvelu.utils.verkkokauppa.payment.types import PaymentStatus
-from tilavarauspalvelu.utils.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
 from utils.sentry import SentryLogger
 
 from tests.factories import OrderFactory, PaymentFactory, PaymentOrderFactory, ReservationFactory
