+++ tilavarauspalvelu/api/graphql/types/merchants/mutations.py
@@ -6,8 +6,8 @@ from graphene_django_extensions.bases import DjangoMutation
 
 from tilavarauspalvelu.api.graphql.extensions.validation_errors import ValidationErrorCodes, ValidationErrorWithCode
 from tilavarauspalvelu.enums import OrderStatus
+from tilavarauspalvelu.integrations.verkkokauppa.payment.exceptions import GetPaymentError
 from tilavarauspalvelu.models import PaymentOrder
-from tilavarauspalvelu.utils.verkkokauppa.payment.exceptions import GetPaymentError
 from utils.sentry import SentryLogger
 
 from .permissions import OrderRefreshPermission
