+++ tilavarauspalvelu/api/graphql/types/reservation/mutations.py
@@ -9,8 +9,8 @@ from graphene_django_extensions import CreateMutation, DeleteMutation, UpdateMut
 
 from tilavarauspalvelu.api.graphql.types.merchants.types import PaymentOrderNode
 from tilavarauspalvelu.enums import OrderStatus, ReservationStateChoice
+from tilavarauspalvelu.integrations.verkkokauppa.order.exceptions import CancelOrderError
 from tilavarauspalvelu.models import Reservation
-from tilavarauspalvelu.utils.verkkokauppa.order.exceptions import CancelOrderError
 
 from .permissions import (
     ReservationCommentPermission,
