+++ tilavarauspalvelu/api/graphql/types/reservation/serializers/confirm_serializers.py
@@ -9,15 +9,18 @@ from tilavarauspalvelu.api.graphql.extensions.validation_errors import Validatio
 from tilavarauspalvelu.api.graphql.types.reservation.serializers.update_serializers import ReservationUpdateSerializer
 from tilavarauspalvelu.enums import Language, OrderStatus, PaymentType, ReservationStateChoice
 from tilavarauspalvelu.integrations.email.main import EmailService
+from tilavarauspalvelu.integrations.verkkokauppa.helpers import (
+    create_mock_verkkokauppa_order,
+    get_verkkokauppa_order_params,
+)
+from tilavarauspalvelu.integrations.verkkokauppa.order.exceptions import CreateOrderError
+from tilavarauspalvelu.integrations.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
 from tilavarauspalvelu.models import PaymentOrder
-from tilavarauspalvelu.utils.verkkokauppa.helpers import create_mock_verkkokauppa_order, get_verkkokauppa_order_params
-from tilavarauspalvelu.utils.verkkokauppa.order.exceptions import CreateOrderError
-from tilavarauspalvelu.utils.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
 from utils.sentry import SentryLogger
 
 if TYPE_CHECKING:
+    from tilavarauspalvelu.integrations.verkkokauppa.order.types import CreateOrderParams, Order
     from tilavarauspalvelu.models import Reservation
-    from tilavarauspalvelu.utils.verkkokauppa.order.types import CreateOrderParams, Order
 
 
 class ReservationConfirmSerializer(ReservationUpdateSerializer):
