+++ tilavarauspalvelu/api/webhooks/views.py
@@ -6,11 +6,11 @@ from rest_framework import viewsets
 from rest_framework.response import Response
 
 from tilavarauspalvelu.enums import OrderStatus
+from tilavarauspalvelu.integrations.verkkokauppa.order.exceptions import GetOrderError
+from tilavarauspalvelu.integrations.verkkokauppa.payment.exceptions import GetPaymentError, GetRefundStatusError
+from tilavarauspalvelu.integrations.verkkokauppa.payment.types import PaymentStatus, RefundStatus
+from tilavarauspalvelu.integrations.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
 from tilavarauspalvelu.models import PaymentOrder
-from tilavarauspalvelu.utils.verkkokauppa.order.exceptions import GetOrderError
-from tilavarauspalvelu.utils.verkkokauppa.payment.exceptions import GetPaymentError, GetRefundStatusError
-from tilavarauspalvelu.utils.verkkokauppa.payment.types import PaymentStatus, RefundStatus
-from tilavarauspalvelu.utils.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
 from utils.date_utils import local_datetime
 from utils.sentry import SentryLogger
