+++ tilavarauspalvelu/integrations/verkkokauppa/constants.py
